# nextjs-frontend/app/openapi-client/schemas.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export const AlbumCreateSchema = {
    properties: {
        title: {
            type: 'string',
            title: 'Title'
        },
        artist: {
            type: 'string',
            title: 'Artist'
        },
        description: {
            anyOf: [
                {
                    type: 'string'
                },
                {
                    type: 'null'
                }
            ],
            title: 'Description'
        }
    },
    type: 'object',
    required: ['title', 'artist', 'description'],
    title: 'AlbumCreate'
} as const;

export const AlbumResponseSchema = {
    properties: {
        title: {
            type: 'string',
            title: 'Title'
        },
        artist: {
            type: 'string',
            title: 'Artist'
        },
        description: {
            anyOf: [
                {
                    type: 'string'
                },
                {
                    type: 'null'
                }
            ],
            title: 'Description'
        },
        id: {
            type: 'integer',
            title: 'Id'
        }
    },
    type: 'object',
    required: ['title', 'artist', 'description', 'id'],
    title: 'AlbumResponse'
} as const;

export const HTTPValidationErrorSchema = {
    properties: {
        detail: {
            items: {
                '$ref': '#/components/schemas/ValidationError'
            },
            type: 'array',
            title: 'Detail'
        }
    },
    type: 'object',
    title: 'HTTPValidationError'
} as const;

export const ValidationErrorSchema = {
    properties: {
        loc: {
            items: {
                anyOf: [
                    {
                        type: 'string'
                    },
                    {
                        type: 'integer'
                    }
                ]
            },
            type: 'array',
            title: 'Location'
        },
        msg: {
            type: 'string',
            title: 'Message'
        },
        type: {
            type: 'string',
            title: 'Error Type'
        }
    },
    type: 'object',
    required: ['loc', 'msg', 'type'],
    title: 'ValidationError'
} as const;