# nextjs-frontend/app/openapi-client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type AlbumCreate = {
    title: string;
    artist: string;
    description: (string | null);
};

export type AlbumResponse = {
    title: string;
    artist: string;
    description: (string | null);
    id: number;
};

export type HTTPValidationError = {
    detail?: Array<ValidationError>;
};

export type ValidationError = {
    loc: Array<(string | number)>;
    msg: string;
    type: string;
};

export type CreateAlbumData = {
    body: AlbumCreate;
};

export type CreateAlbumResponse = (AlbumResponse);

export type CreateAlbumError = (HTTPValidationError);

export type ReadAlbumsData = {
    query?: {
        limit?: number;
        skip?: number;
    };
};

export type ReadAlbumsResponse = (Array<AlbumResponse>);

export type ReadAlbumsError = (HTTPValidationError);